Implement Warning the user after successful added if the series was 100 before to reach the maximum 10 digits SeriesNumber in CR

+++ Models/CollectionReceipt.cs
@@ -12,7 +12,7 @@ namespace Accounting_System.Models
         public SalesInvoice? SalesInvoice { get; set; }
 
         public string? CRNo { get; set; }
-        public int SeriesNumber { get; set; }
+        public long SeriesNumber { get; set; }
 
         public DateTime Date { get; set; }
